requirements.txt
asgiref==3.3.4
boto3==1.17.105
botocore==1.20.105
dj-database-url==0.5.0
Django==3.2.3
django-environ==0.4.5
django-heroku==0.3.1
django-ipware==3.0.2
django-storages==1.11.1
django-user-agents==0.4.0
gunicorn==20.1.0
html2image==1.1.2
jmespath==0.10.0
mysqlclient==2.0.3
numpy==1.21.0
pandas==1.2.4
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
whitenoise==5.2.0
